pyproject.toml
[project]
name = "langmem-tutorials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=0.2.5",
    "langgraph>=0.3.8",
    "langmem>=0.0.15",
    "python-dotenv>=1.0.1",
]
